OSIS-2290 Move check_acronym test
